requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.50.0
gspread==6.1.2
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
python-dotenv==1.0.1
requests==2.32.3
python-telegram-bot==21.6
